Port torpedo.rake to openSUSE

+++ tasks/torpedo.rake
@@ -25,10 +25,13 @@ task :torpedo do
 	remote_exec %{
 ssh #{server_name} bash <<-"EOF_SERVER_NAME"
 
-#TODO-vuntz
 if [ -f /bin/rpm ]; then
-	rpm -q rubygems &> /dev/null || yum install -y rubygems &> /dev/null
-	rpm -q rubygem-json &> /dev/null || yum install -y rubygem-json &> /dev/null
+	if [ -f /etc/fedora-release ]; then
+		rpm -q rubygems &> /dev/null || yum install -y rubygems &> /dev/null
+		rpm -q rubygem-json &> /dev/null || yum install -y rubygem-json &> /dev/null
+	else
+		rpm -q rubygem-json &> /dev/null || zypper -q --non-interactive install rubygem-json &> /dev/null
+	fi
 	if ruby -v | grep 1\.9\. &> /dev/null; then
 		gem install --no-rdoc --no-ri test-unit
 	fi
